+++ netuitive/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 __author__ = 'Netuitive, Inc'
-__version__ = '0.0.9'
+__version__ = '0.1.0'
 
 from .client import Client  # nopep8
 from .element import Element  # nopep8